# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flight Booking Management API powered by Duffel
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type SearchOffersBodyCabinClass =
  (typeof SearchOffersBodyCabinClass)[keyof typeof SearchOffersBodyCabinClass];

export const SearchOffersBodyCabinClass = {
  economy: "economy",
  premium_economy: "premium_economy",
  business: "business",
  first: "first",
} as const;

export type PassengerInputType =
  (typeof PassengerInputType)[keyof typeof PassengerInputType];

export const PassengerInputType = {
  adult: "adult",
  child: "child",
  infant_without_seat: "infant_without_seat",
} as const;

export interface PassengerInput {
  type: PassengerInputType;
  age?: number;
}

export interface SearchOffersBody {
  /** IATA airport code for origin */
  origin: string;
  /** IATA airport code for destination */
  destination: string;
  /** Departure date (YYYY-MM-DD) */
  departureDate: string;
  /** Return date for round trips (YYYY-MM-DD) */
  returnDate?: string;
  passengers: PassengerInput[];
  cabinClass?: SearchOffersBodyCabinClass;
}

export type SearchOffersResponseSlicesItem = { [key: string]: unknown };

export interface Airport {
  iataCode: string;
  name: string;
  cityName?: string;
  countryName?: string;
}

export interface Airline {
  iataCode: string;
  name: string;
  logoSymbolUrl?: string;
  logotypeLockupImageUrl?: string;
}

export type BaggageAllowanceType =
  (typeof BaggageAllowanceType)[keyof typeof BaggageAllowanceType];

export const BaggageAllowanceType = {
  carry_on: "carry_on",
  checked: "checked",
} as const;

export interface BaggageAllowance {
  type: BaggageAllowanceType;
  quantity: number;
}

export type SegmentAircraft = {
  iataCode?: string;
  name?: string;
};

export interface Segment {
  id: string;
  origin: Airport;
  destination: Airport;
  departureDateTime: string;
  arrivalDateTime: string;
  duration?: string;
  flightNumber?: string;
  marketingCarrier?: Airline;
  operatingCarrier?: Airline;
  aircraft?: SegmentAircraft;
  /** Baggage allowances for the first passenger on this segment */
  baggages?: BaggageAllowance[];
}

export interface Slice {
  id: string;
  origin: Airport;
  destination: Airport;
  departureDateTime: string;
  arrivalDateTime: string;
  duration?: string;
  segments?: Segment[];
}

export type OfferPassengersItem = { [key: string]: unknown };

export interface Offer {
  id: string;
  totalAmount: string;
  totalCurrency: string;
  baseAmount?: string;
  taxAmount?: string;
  expiresAt?: string;
  cabinClass?: string;
  slices: Slice[];
  passengers?: OfferPassengersItem[];
  owner?: Airline;
}

export interface SearchOffersResponse {
  offerRequestId: string;
  offers: Offer[];
  slices?: SearchOffersResponseSlicesItem[];
}

export type CreateOrderBodyType =
  (typeof CreateOrderBodyType)[keyof typeof CreateOrderBodyType];

export const CreateOrderBodyType = {
  instant: "instant",
  hold: "hold",
} as const;

export type PassengerDetailsTitle =
  (typeof PassengerDetailsTitle)[keyof typeof PassengerDetailsTitle];

export const PassengerDetailsTitle = {
  mr: "mr",
  ms: "ms",
  mrs: "mrs",
  miss: "miss",
  dr: "dr",
} as const;

export type PassengerDetailsGender =
  (typeof PassengerDetailsGender)[keyof typeof PassengerDetailsGender];

export const PassengerDetailsGender = {
  m: "m",
  f: "f",
} as const;

export interface PassengerDetails {
  id: string;
  title?: PassengerDetailsTitle;
  givenName: string;
  familyName: string;
  gender: PassengerDetailsGender;
  bornOn: string;
  email: string;
  phoneNumber: string;
  passportNumber?: string;
  passportExpiryDate?: string;
  nationalityIataCountryCode?: string;
}

export interface CreateOrderBody {
  selectedOfferId: string;
  passengers: PassengerDetails[];
  type?: CreateOrderBodyType;
}

export type OrderPassengersItem = { [key: string]: unknown };

export type OrderDocumentsItem = { [key: string]: unknown };

export type OrderPaymentStatus = {
  awaitingPayment?: boolean;
  paymentRequiredBy?: string;
};

export interface Order {
  id: string;
  bookingReference: string;
  status: string;
  totalAmount: string;
  totalCurrency: string;
  createdAt: string;
  passengers?: OrderPassengersItem[];
  slices?: Slice[];
  owner?: Airline;
  documents?: OrderDocumentsItem[];
  paymentStatus?: OrderPaymentStatus;
}

export type ListOrdersResponseMeta = {
  limit?: number;
  before?: string;
  after?: string;
};

export interface ListOrdersResponse {
  orders: Order[];
  meta: ListOrdersResponseMeta;
}

export interface CancellationResponse {
  id: string;
  orderId: string;
  refundAmount?: string;
  refundCurrency?: string;
  refundTo?: string;
  confirmedAt?: string;
  expiresAt?: string;
}

export type StatsSummaryTopRoutesItem = {
  origin?: string;
  destination?: string;
  count?: number;
};

export interface StatsSummary {
  totalOrders: number;
  confirmedOrders: number;
  cancelledOrders: number;
  totalRevenue: string;
  currency: string;
  recentOrders: Order[];
  topRoutes: StatsSummaryTopRoutesItem[];
}

export type ListOrdersParams = {
  limit?: number;
  after?: string;
  before?: string;
};