# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flight Booking Management API powered by Duffel
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers/search:
    post:
      operationId: searchOffers
      tags: [offers]
      summary: Search for flight offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchOffersBody"
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchOffersResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers/{offerId}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get a specific offer by ID
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SearchOffersBody:
      type: object
      properties:
        origin:
          type: string
          description: IATA airport code for origin
        destination:
          type: string
          description: IATA airport code for destination
        departureDate:
          type: string
          format: date
          description: Departure date (YYYY-MM-DD)
        returnDate:
          type: string
          format: date
          description: Return date for round trips (YYYY-MM-DD)
        passengers:
          type: array
          items:
            $ref: "#/components/schemas/PassengerInput"
        cabinClass:
          type: string
          enum: [economy, premium_economy, business, first]
          default: economy
      required:
        - origin
        - destination
        - departureDate
        - passengers
    PassengerInput:
      type: object
      properties:
        type:
          type: string
          enum: [adult, child, infant_without_seat]
        age:
          type: integer
      required:
        - type
    SearchOffersResponse:
      type: object
      properties:
        offerRequestId:
          type: string
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        slices:
          type: array
          items:
            type: object
      required:
        - offerRequestId
        - offers
    Offer:
      type: object
      properties:
        id:
          type: string
        totalAmount:
          type: string
        totalCurrency:
          type: string
        baseAmount:
          type: string
        taxAmount:
          type: string
        expiresAt:
          type: string
        cabinClass:
          type: string
        slices:
          type: array
          items:
            $ref: "#/components/schemas/Slice"
        passengers:
          type: array
          items:
            type: object
        owner:
          $ref: "#/components/schemas/Airline"
      required:
        - id
        - totalAmount
        - totalCurrency
        - slices
    Slice:
      type: object
      properties:
        id:
          type: string
        origin:
          $ref: "#/components/schemas/Airport"
        destination:
          $ref: "#/components/schemas/Airport"
        departureDateTime:
          type: string
        arrivalDateTime:
          type: string
        duration:
          type: string
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
      required:
        - id
        - origin
        - destination
        - departureDateTime
        - arrivalDateTime
    BaggageAllowance:
      type: object
      properties:
        type:
          type: string
          enum: [carry_on, checked]
        quantity:
          type: integer
      required:
        - type
        - quantity
    Segment:
      type: object
      properties:
        id:
          type: string
        origin:
          $ref: "#/components/schemas/Airport"
        destination:
          $ref: "#/components/schemas/Airport"
        departureDateTime:
          type: string
        arrivalDateTime:
          type: string
        duration:
          type: string
        flightNumber:
          type: string
        marketingCarrier:
          $ref: "#/components/schemas/Airline"
        operatingCarrier:
          $ref: "#/components/schemas/Airline"
        aircraft:
          type: object
          properties:
            iataCode:
              type: string
            name:
              type: string
        baggages:
          type: array
          description: Baggage allowances for the first passenger on this segment
          items:
            $ref: "#/components/schemas/BaggageAllowance"
      required:
        - id
        - origin
        - destination
        - departureDateTime
        - arrivalDateTime
    Airport:
      type: object
      properties:
        iataCode:
          type: string
        name:
          type: string
        cityName:
          type: string
        countryName:
          type: string
      required:
        - iataCode
        - name
    Airline:
      type: object
      properties:
        iataCode:
          type: string
        name:
          type: string
        logoSymbolUrl:
          type: string
        logotypeLockupImageUrl:
          type: string
      required:
        - iataCode
        - name
...