# lib/api-zod/src/generated/types/baggageAllowance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flight Booking Management API powered by Duffel
 * OpenAPI spec version: 0.1.0
 */
import type { BaggageAllowanceType } from "./baggageAllowanceType";

export interface BaggageAllowance {
  type: BaggageAllowanceType;
  quantity: number;
}