# lib/api-zod/src/generated/types/segment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flight Booking Management API powered by Duffel
 * OpenAPI spec version: 0.1.0
 */
import type { Airline } from "./airline";
import type { Airport } from "./airport";
import type { BaggageAllowance } from "./baggageAllowance";
import type { SegmentAircraft } from "./segmentAircraft";

export interface Segment {
  id: string;
  origin: Airport;
  destination: Airport;
  departureDateTime: string;
  arrivalDateTime: string;
  duration?: string;
  flightNumber?: string;
  marketingCarrier?: Airline;
  operatingCarrier?: Airline;
  aircraft?: SegmentAircraft;
  /** Baggage allowances for the first passenger on this segment */
  baggages?: BaggageAllowance[];
}